separate Tag and TagFactory
